+++ src/dyscord/client/discord_client.py
@@ -136,15 +136,24 @@ class DiscordClient:
             self.intent += Intents.DIRECT_MESSAGE_TYPING
         self._intents_defined = True
 
-    def run(self):
-        '''Start the async loop and run forever.'''
+    def run(self, loop = None):
+        '''Start the async loop and run forever.
+        Arguments:
+            loop: If desired, use a given asyncio compatible loop. One will be created if not given.
+        '''
         self._log.info('Starting...')
         self._log.info(f'Application ID: [{self.application_id}]')
         self._log.info(f'Version: [{__version__}]')
-        nest_asyncio.apply()
-        asyncio.run(self._run())
 
-    async def _run(self):
+        loop = loop if loop is not None else asyncio.get_event_loop()
+
+        nest_asyncio.apply(loop)
+
+        loop.create_task(self._run())
+
+        loop.run_forever()
+
+    async def _run(self, ):
 
         if self._intents_defined is False:
             warnings.warn('Started without defining intents. Client will likely get ZERO input. Consider calling the \'configure_intents\' function.', UserWarning)