+++ docs/changelog.md
@@ -2,7 +2,11 @@
 
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.3.0] WIP
+## [v0.4.0] 2021-10-22
+### Added
+- Handle invalid session events.
+
+## [v0.3.0] 2021-10-22
 ### Added
 - Built in reconnect support.
 - Support giving an event loop to DiscordClient.run()